feat: add appropriate meta data for each page
