+++ src/utils/measureExecutionTime.ts
@@ -1,0 +1,11 @@
+import { performance } from "perf_hooks";
+
+export const measureExecutionTime = (func: any) => {
+  const startTime = performance.now();
+  func();
+  const endTime = performance.now();
+
+  const timeTaken = endTime - startTime;
+
+  return timeTaken;
+};
